+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # v2.11.1
 
 * added extra operators to `Predicate`
+* deprecated `skip_first` and `stop_after`